This is the initial commit for the qgis plugin template. The template consists of:
class Plugin (a subclass of QgisPlugin)
class PluginGuiBase (supplied as a simple qdialog ui)
class PluginGui (the subclass of PluginGuiBase)

The only thing the user would need to change in the makefile is the name of the .so
so you just copy the plugin template dir to a new dir under plugins, change the .so name in the makefile add the plugin dir the configure.in and the plugin dir makefile.

I will be adding a bash script file soon that automates the above process. This template development is in progress.


git-svn-id: http://svn.osgeo.org/qgis@881 c8812cc2-4d05-0410-92ff-de0c093fc19c