QgsProject::read() now kinda supports reading project files with raster layers.  The legend doesn't seem to properly restore, but I suspect that's largely because certain signals and slots aren't being connected.
Added QgsRasterLayer::readXML_() which contains code previously found in
QgsProjectIO::read().

Simplified QgsProject::read() a bit.

Added QgsRasterLayer::readFile(); that which was in its ctor resposible for
reading files has been moved to that function.  This allows for explicitly
loading rasters when the ctor had previously been invoked with null
arguments.

Note that this CVS commit is on the qgsproject-branch.


git-svn-id: http://svn.osgeo.org/qgis@2017 c8812cc2-4d05-0410-92ff-de0c093fc19c