Fix for clipping of labels when labelling polygon features.
Modified label rendering procedure so that it is a seprate call (drawLabels) and no longer part of the draw() method. Added a second loop in QgsMapCanvas::render() to go through all visible layers and call their drawLabels method. This loop is called after the features rendereing loop so the clipping problem is thusly forthwithly avaoided. Performance doesnt seem to bad either.


git-svn-id: http://svn.osgeo.org/qgis@2024 c8812cc2-4d05-0410-92ff-de0c093fc19c