Almost completed writing vector layer info.  Still have to address writing renderer and label XML.
Almost have comleted kludgey cheat for renderer XML.  Currently renderers use
ofstreams to write their XML to directly.  The "new" way would be to add DOM
nodes.  However in the interest of time, I'm hoping to take the XML stream
from the renderer, read it into a QDomDocument, and then import that into the
main document.  Have basic mechanism in place to do that, but am getting
import errors.  Probably have to add an XML header to get import to work
correctly.


git-svn-id: http://svn.osgeo.org/qgis@2060 c8812cc2-4d05-0410-92ff-de0c093fc19c