Fixed bug #1040377, whereby user wouldn't be prompted to save current project if changes had been made while creating a new project or opening an existing one.
qgis is a bit better of setting project dirty bit when a change has been made
since it's now set in more places than in the original code.  (E.g., panning
and zooming through the canvas, setting layer visibility, adding or removing
from overview, etc.)  However, I'm not completely confident I set the dirty
bit for all state changes.  In the future, when (or if) an undo/re-do is
implemented using a Command design pattern, the project dirty bit management
should be consolidated in the Command objects.


git-svn-id: http://svn.osgeo.org/qgis@2104 c8812cc2-4d05-0410-92ff-de0c093fc19c