Implemented fix for bug #1077208.  I.e., the following logic implemented for application title bar text:
  If the current project title is null
     if the project file is null then
         set title text to just application name and version
     else
         set set title text to the the project file name
  else
     set the title text to project title

Note that the behavior of setting a "modified" flag or text on the title bar
when the project dirty bit is set hasn't been implemented.  Not only should
that be a separate feature request, but will take a little more work to
implement.  It could also be argued that it's unnecessary since just about any
user actions will make the project dirty, even just panning and zooming since
the AOI is saved and restored in projects.


git-svn-id: http://svn.osgeo.org/qgis@2411 c8812cc2-4d05-0410-92ff-de0c093fc19c