fix for bug [ 1077211 ] Map units not preseved
Actually the bug manifested itself in other ways too : if I saved a project in units degrees, then created a new blank project, the props dialog for the project showed the units for the new project as degrees when in fact the units were meters. This is all fixed now.


git-svn-id: http://svn.osgeo.org/qgis@2423 c8812cc2-4d05-0410-92ff-de0c093fc19c