Added srid column to the listview
Changed the QMap that stored WKT of a projection to store the SPATIAL_REF_SYS
structure for each projection. The structure contains the srid, wkt text, and
proj4 parameters (among other items). This expands the list of projections
available to over 1,700.

Projection support is currently broken until the changes from storing WKT to
a SPATIAL_REF_SYS struct are completed (not that it really worked anyway)


git-svn-id: http://svn.osgeo.org/qgis/branches/Projections_Branch@2664 c8812cc2-4d05-0410-92ff-de0c093fc19c