Fix for bug #1285012 whereby multiple instances of file filters would show up after project file loaded and new layers were added.
Note, though, that the "finding missing layers" implementation, which
admittedly has nothing to do with this bug, only worked if the project file
was loaded from the command line.  Now that functionality works for projects
loaded from the GUI.  Unfortunately there needs to be some tweaking since it's
hardcoded to use vector format dialogs; it doesn't intelligently handle
vectors vs. rasters depending on what files have gone missing.  And it doesn't
appear to properly populate the file dialogs with the missing file name.


git-svn-id: http://svn.osgeo.org/qgis@3895 c8812cc2-4d05-0410-92ff-de0c093fc19c