The map should now move again when you go to drag it with the little hand/pan symbol, not just when you complete the drag.
The original problem was due to the bitBlt idiom going away in Qt4.

Others should check that this new method works/doesn't cause flickering under their system, as I developed this modification against a KDE session via VNC.

Also there seems to be a little bit of realignment (probabaly a fraction of a screen pixel) that occurs when the pan completes.  I presume this is not a significant issue to anyone.


git-svn-id: http://svn.osgeo.org/qgis@4283 c8812cc2-4d05-0410-92ff-de0c093fc19c