Fix for ticket #111.
Changed the code from using text to manipulate the scale bar and north
arrow to using ints as this is more robust, especially when the user
changes languages and saves a project.

Note: this commit will break saved project files if they include scale
bar or north arrow sections (the scale bar and north arrow will end up
in the default locations rather than those when the project was saved).



git-svn-id: http://svn.osgeo.org/qgis@5444 c8812cc2-4d05-0410-92ff-de0c093fc19c