In the Attributes Table, the Copy button is now hooked up to the same editCopy() procedure that the main application uses.  The copySelectedRowsToClipboard() procedure that was used in QgsAttributeTableDisplay is no longer referenced.
As part of this "merge", the main application copy will now precede its results with a row of attribute headers.

Also, the editCopy() procedure and friends (editCut() and editPaste()) can now take the layer to act on, thus allowing the Attribute Table to work on a different layer to the Current Layer in the Legend.

This change was done so that the Copy button in the main application AND the Copy button on the attribute table works consistently.



git-svn-id: http://svn.osgeo.org/qgis@5584 c8812cc2-4d05-0410-92ff-de0c093fc19c