When Stopping Editing and Saving Changes, if the commit was *unsuccessful*, retain the in-memory edits.  Previously the in-memory edits would be erased anyway.
Note this is not a complete solution as commits are currently a 3- or 4-stage process, and if one stage fails the chain cannot always be rolled-back easily.  If this is the case then a detailed error message is given to the user to assist their understanding of the discrepency.

This commit should be a bug fix for trac #131, but I don't have the environment to test it fully - the bug submitter will have to assist there.

Bonus feature: Tidied up some variable names in QgsVectorLayer - makes it clearer what they contain.



git-svn-id: http://svn.osgeo.org/qgis@5591 c8812cc2-4d05-0410-92ff-de0c093fc19c