When loading a project file, restore the canvas extent to that provided in the project file (i.e. the extent in existence when the project file was saved).  Previously the extent was the maximum extent across all layers.
I'm not sure why this functionality hasn't been done before, as it seemed so trivial to implement?

The main benefit to its implementation is that in large layers, the user doesn't have to wait for the whole layer to draw before he can restart work on a project (which is particularly important since we currently can no longer hit Esc to halt a redraw).



git-svn-id: http://svn.osgeo.org/qgis@5613 c8812cc2-4d05-0410-92ff-de0c093fc19c